+++ CHANGELOG.md
@@ -1,15 +1,52 @@
+## [1.5.0]
+* printlnMixedKanji merged to println
+* Can print rows/cols with mixed chinese/non-chinese texts
+
+
+## [1.4.1]
+* Print mixed (chinese + latin) text. Only for printers supporting Kanji mode
+
+
+## [1.4.0]
+* Print barcodes
+
+
+## [1.3.2]
+* Added alternative printImageRaster method using (GS v 0) obsolete command
+* Code refactoring
+
+
+## [1.3.1]
+* Removed raw lib
+* Code refactoring
+
+
+## [1.3.0]
+* Image printing method has been improved
+* Using ESC * command to print images instead of outdated GS v 0
+* Added print image Flutter example
+
+
+## [1.2.0]
+* Printing images (beta)
+  
+
 ## [1.1.3]
 * printCodeTable bug fixed
-* updated test print example 
+* Updated test print example 
+
 
 ## [1.1.2]
 * Better alignment
 
+
 ## [1.1.1]
 * Fixed columns alignment bug
 
+
 ## [1.1.0]
 * Added code page support
+
 
 ## [1.0.2]
 * Send raw command(s)